Split feature overlay manager and service
